refactor: 구입 금액을 PurchaseAmount로 포장
