+++ src/main/java/domain/PurchaseAmount.java
@@ -1,0 +1,28 @@
+package domain;
+
+public class PurchaseAmount {
+    private static final int LOTTO_PRICE = 1000;
+    private final int amount;
+
+    public PurchaseAmount(int amount) {
+        validateAmount(amount);
+        this.amount = amount;
+    }
+
+    private void validateAmount(int amount) {
+        if (amount < LOTTO_PRICE) {
+            throw new IllegalArgumentException("구입 금액은 1000원 이상이어야 합니다.");
+        }
+        if (amount % LOTTO_PRICE != 0) {
+            throw new IllegalArgumentException("구입 금액은 1000원 단위여야 합니다.");
+        }
+    }
+
+    public int calculateLottoCount() {
+        return amount / LOTTO_PRICE;
+    }
+
+    public int getAmount() {
+        return amount;
+    }
+}
